+++ seed-electron-react-basic/jest.config.js
@@ -1,0 +1,13 @@
+module.exports = {
+  testPathIgnorePatterns: [
+    '<rootDir>/node_modules/',
+    '<rootDir>/.vscode',
+    '<rootDir>/dist',
+    '<rootDir>/build',
+    '<rootDir>/.next',
+    '<rootDir>/scripts',
+    '<rootDir>/pages',
+    '<rootDir>/doc',
+    '<rootDir>/typings',
+  ],
+}